refactor generator
